stringify
It turns out concatenating strings isn't the *worst* thing to do, and
streams can be pretty expensive. This is pretty much a total rewrite,
optimizing for rendering to a single string. This turns out to be not
too bad, performance-wise, for the streaming case too, by doing a single
write.

`deep-access` was re-written and in-lined for performance.

Added `asString` function to deal with string rendering directly. In
v2.0.0, this will probably be only way to use `maudite`.